move 8bit vector methods into util.hnsw
