+++ lucene/core/src/java/org/apache/lucene/util/VectorUtil.java
@@ -213,21 +213,4 @@ public final class VectorUtil {
       u[i] += v[i];
     }
   }
-
-  public static float dotProduct(BytesRef a, int aOffset, BytesRef b, int bOffset, int len) {
-    // fixme -- move to codec? What if later we want to access the bytes some other way?
-    int total = 0;
-    for (int i = 0; i < len; i++) {
-      total += a.bytes[aOffset++] * b.bytes[bOffset++];
-    }
-    return total;
-  }
-
-  public static BytesRef toBytesRef(float[] vector) {
-    BytesRef b = new BytesRef(vector.length);
-    for (int i = 0; i < vector.length; i++) {
-      b.bytes[i] = (byte) vector[i];
-    }
-    return b;
-  }
 }
